pyproject.toml
[tool.poetry]
name = "user_be"
version = "1.0.0"
description = "Shoper project"
authors = ["SerhiiLakodei"]

[tool.poetry.dependencies]
python = "^3.11.6"
fastapi = "^0.110.0"
python-json-logger = "^2.0.7"
sentry-sdk = "^1.43.0"
uvicorn = "^0.29.0"
uvloop = "^0.19.0"
gunicorn = "^21.2.0"
ujson = "^5.9.0"
pydantic = {extras = ["email"], version = "^2.6.4"}
pydantic-settings = "^2.2.1"
aioboto3 = "^12.3.0"
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
cryptography = "^42.0.5"
beanie = "^1.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
